feat: add deleteContactRouter

+++ src/server/controller/deleteFriendController.js
@@ -1,15 +1,22 @@
 import deleteContactService from "../../services/deleteContactService.js";
 
-class deleteContactController {
+class DeleteContactController {
     async removeFriend(req, res) {
-        const { contactId } = req.params;
-        const currentContactId = req.contact.id;
-
         try {
+            const { contactId } = req.params;
+            const currentContactId = req.user.id;
+
+            if (!currentContactId) {
+                return res
+                    .status(401)
+                    .json({ error: "Usuário não autenticado." });
+            }
+
             const response = await deleteContactService.removeFriend(
                 currentContactId,
                 parseInt(contactId),
             );
+
             res.status(200).json(response);
         } catch (error) {
             res.status(400).json({ error: error.message });
@@ -17,4 +24,4 @@ class deleteContactController {
     }
 }
 
-export default new deleteContactController();
+export default new DeleteContactController();
